requirements.txt
Django==2.2.18
psycopg2-binary==2.8.4
Pillow
coverage==4.5.2
djangorestframework
requests
